change able
Right I am trying to get it so that when it says the password it will ask if it is up to date and if you say no you would be able to change it. but that requires me putting all of the passwords into sepurate variables in order to change them.